fix(foundation): merge identity is the canonical ref; user order and per-skill fields win
Cross-model review found seven defects, one a blocker:

1. [BLOCKER] refOf folded the alias into identity, so a curated skill the user
   exposed under another name was reported missing AND extra, and the merge
   emitted BOTH objects — the same skill twice in one preset. Identity for
   merging is now the bare ref; an entry on both sides keeps the USER'S object.
2. The merge rebuilt from the shipped array, reverting the order the user
   arranged and dropping per-skill fields they set. Now the stored array is
   preserved verbatim and only genuinely new refs are appended. This reverses
   the original 'curated order first' rule: a reordered preset with lost edits
   is a worse failure than a new skill landing last.
3. doctor: pending = updatable + added made the customized branch unreachable,
   hiding customized entries that also needed adopting. Total now includes it.
4. A 'new' preset kept the curated constant's fixed epoch timestamps; both are
   stamped from one now() read.
5. `foundation <id>` ignored the filter without --adopt.
6. detectPlanDrift interpolated paths uncapped into evidence kept in telemetry.
7. changedFields omitted 'color', so recolouring misclassified as updatable.

diffSkills derives both sets from the ref alone; refOf survives as the display
label, so output still reads 'a/x as my-alias' while an alias alone can no
longer flip an entry to customized.

Conductor verified: tsc 0, build 0, 125/125, plus a direct run of the blocker
case — one entry for the aliased ref, alias and whenToUse intact, user order
kept, new skill appended, second adopt a no-op.

+++ src/host/practices/detectors.ts
@@ -221,7 +221,9 @@ export function detectPlanDrift(view: SessionView): PracticeResult {
   const drift = view.drift ?? []
   if (drift.length === 0) return result('plan-drift', 'green', ['every edit is named in plan.md'])
   if (view.planUpdated === true) return result('plan-drift', 'green', [`plan.md updated after ${drift.length} unplanned edit${drift.length === 1 ? '' : 's'}`])
-  return result('plan-drift', 'amber', [`${drift.length} file${drift.length === 1 ? '' : 's'} not in plan.md: ${drift.slice(0, 3).map(d => d.path).join(', ')}`, 'update plan.md in the same branch, or say why'], drift[0].t)
+  // Raw evidence is persisted to telemetry and the scorecard, not only
+  // rendered into the prompt, so the cap has to happen here as well.
+  return result('plan-drift', 'amber', [`${drift.length} file${drift.length === 1 ? '' : 's'} not in plan.md: ${short(drift.slice(0, 3).map(d => d.path).join(', '))}`, 'update plan.md in the same branch, or say why'], drift[0].t)
 }
 
 export function detectWorktreeHygiene(view: SessionView): PracticeResult {
